Splits the C++ builders into different headers
