+++ libs/framework/include/celix/TrackerBuilders.h
@@ -19,311 +19,13 @@
 
 #pragma once
 
+#include <string>
+#include <vector>
 #include <memory>
-#include <functional>
 
-#include "celix/ServiceRegistration.h"
-#include "celix/Trackers.h"
+#include "Trackers.h"
 
 namespace celix {
-
-    /**
-     * Fluent builder API to build a new service registration for a service of type I.
-     *
-     * \warning Not thread safe.
-     * @tparam I The service type (Note should be the abstract interface, not the interface implementer)
-     */
-    template<typename I>
-    class ServiceRegistrationBuilder {
-    private:
-        friend class BundleContext;
-
-        //NOTE private to prevent move so that a build() call cannot be forgotten
-        ServiceRegistrationBuilder(ServiceRegistrationBuilder&&) = default;
-    public:
-        ServiceRegistrationBuilder(std::shared_ptr<celix_bundle_context_t> _cCtx, std::shared_ptr<I> _svc, std::string _name) :
-                cCtx{std::move(_cCtx)},
-                svc{std::move(_svc)},
-                name{std::move(_name)} {
-
-        }
-
-        ServiceRegistrationBuilder& operator=(ServiceRegistrationBuilder&&) = delete;
-        ServiceRegistrationBuilder(const ServiceRegistrationBuilder&) = delete;
-        ServiceRegistrationBuilder operator=(const ServiceRegistrationBuilder&) = delete;
-
-        /**
-         * Set the service version.
-         * This will lead to a 'service.version' service property.
-         */
-        ServiceRegistrationBuilder& setVersion(std::string v) { version = std::move(v); return *this; }
-
-        /**
-         * Add a property to the service properties.
-         * If a key is already present the value will be overridden.
-         */
-        ServiceRegistrationBuilder& addProperty(std::string key, std::string value) { properties.set(std::move(key), std::move(value)); return *this; }
-
-        /**
-         * Add a property to the service properties.
-         * If a key is already present the value will be overridden.
-         */
-        ServiceRegistrationBuilder& addProperty(std::string key, const char* value) { properties.set(std::move(key), std::string{value}); return *this; }
-
-        /**
-         * Add a property to the service properties.
-         * If a key is already present the value will be overridden.
-         */
-        template<typename T>
-        ServiceRegistrationBuilder& addProperty(std::string key, T value) { properties.set(std::move(key), std::to_string(std::move(value))); return *this; }
-
-        /**
-         * Set the service properties.
-         * Note this call will clear any already set properties.
-         */
-        ServiceRegistrationBuilder& setProperties(celix::Properties p) { properties = std::move(p); return *this; }
-
-        /**
-         * Add the properties to the service properties.
-         * Note this call will add these properties to the already set properties.
-         * If a key is already present the value will be overridden.
-         */
-        ServiceRegistrationBuilder& addProperties(const celix::Properties& props) {
-            for (const auto& pair : props) {
-                properties.set(pair.first, pair.second);
-            }
-            return *this;
-        }
-
-        /**
-         * Adds an on registered callback for the service registration.
-         * This callback will be called on the Celix event thread when the service is registered (REGISTERED state)
-         */
-        ServiceRegistrationBuilder& addOnRegistered(std::function<void(ServiceRegistration&)> callback) {
-            onRegisteredCallbacks.emplace_back(std::move(callback));
-            return *this;
-        }
-
-        /**
-         * Adds an on unregistered callback for the service registration.
-         * This callback will be called on the Celix event thread when the service is unregistered (UNREGISTERED state)
-         */
-        ServiceRegistrationBuilder& addOnUnregistered(std::function<void(ServiceRegistration&)> callback) {
-            onUnregisteredCallbacks.emplace_back(std::move(callback));
-            return *this;
-        }
-
-        /**
-         * Configure if the service registration will be done synchronized or asynchronized.
-         * When a service registration is done synchronized the underlining service will be registered in the Celix
-         * framework (and all service trackers will have been informed) when the build() returns.
-         * For asynchronized the build() will not wait until this is done.
-         *
-         * The benefit of synchronized is the user is ensured a service registration is completely done, but extra
-         * care has te be taking into account to prevent deadlocks:
-         * lock object -> register service -> trigger service tracker -> trying to update a locked object.
-         *
-         * The benefit of asynchronized is that service registration can safely be done inside a lock.
-         *
-         * Default behavior is asynchronized.
-         */
-        ServiceRegistrationBuilder& setRegisterAsync(bool async) {
-            registerAsync = async;
-            return *this;
-        }
-
-        /**
-         * Configure if the service un-registration will be done synchronized or asynchronized.
-         * When a service un-registration is done synchronized the underlining service will be unregistered in the Celix
-         * framework (and all service trackers will have been informed) when the ServiceRegration::unregister is
-         * called of when the ServiceRegistration goes out of scope (with a still REGISTERED service).
-         * For asynchronized this will be done asynchronously and the actual ServiceRegistration object will only be
-         * deleted if the un-registration is completed async.
-         *
-         * The benefit of synchronized is the user is ensured a service un-registration is completely done when
-         * calling unregister() / letting ServiceRegistration out of scope, but extra
-         * care has te be taking into account to prevent deadlocks:
-         * lock object -> unregister service -> trigger service tracker -> trying to update a locked object.
-         *
-         * The benefit of asynchronized is that service un-registration can safely be called inside a lock.
-         *
-         * Default behavior is asynchronized.
-         */
-        ServiceRegistrationBuilder& setUnregisterAsync(bool async) {
-            unregisterAsync = async;
-            return *this;
-        }
-
-        /**
-         * "Builds" the service registration and return a ServiceRegistration.
-         *
-         * The ServiceRegistration will register the service to the Celix framework and unregister the service if
-         * the shared ptr for the ServiceRegistration goes out of scope.
-         *
-         */
-        std::shared_ptr<ServiceRegistration> build() {
-            return ServiceRegistration::create(
-                    cCtx,
-                    std::move(svc),
-                    std::move(name),
-                    std::move(version),
-                    std::move(properties),
-                    registerAsync,
-                    unregisterAsync,
-                    std::move(onRegisteredCallbacks),
-                    std::move(onUnregisteredCallbacks));
-        }
-    private:
-        const std::shared_ptr<celix_bundle_context_t> cCtx;
-        std::shared_ptr<I> svc;
-        std::string name;
-        std::string version{};
-        celix::Properties properties{};
-        bool registerAsync{true};
-        bool unregisterAsync{true};
-        std::vector<std::function<void(ServiceRegistration&)>> onRegisteredCallbacks{};
-        std::vector<std::function<void(ServiceRegistration&)>> onUnregisteredCallbacks{};
-    };
-
-    /**
-     * Fluent builder API to use a service(s) of type I.
-     *
-     * \warning Not thread safe.
-     * @tparam I The service type to use
-     */
-    template<typename I>
-    class UseServiceBuilder {
-    private:
-        friend class BundleContext;
-
-        //NOTE private to prevent move so that a build() call cannot be forgotten
-        UseServiceBuilder(UseServiceBuilder&&) = default;
-    public:
-        explicit UseServiceBuilder(std::shared_ptr<celix_bundle_context_t> _cCtx, std::string _name, bool _useSingleService = true) :
-                cCtx{std::move(_cCtx)},
-                name{std::move(_name)},
-                useSingleService{_useSingleService} {
-        }
-
-        UseServiceBuilder& operator=(UseServiceBuilder&&) = delete;
-        UseServiceBuilder(const UseServiceBuilder&) = delete;
-        UseServiceBuilder operator=(const UseServiceBuilder&) = delete;
-
-        /**
-         * Set filter to be used to select a service.
-         * The filter must be LDAP filter.
-         * Example: "(property_key=value)"
-         */
-        UseServiceBuilder& setFilter(std::string f) { filter = celix::Filter{std::move(f)}; return *this; }
-
-        /**
-         * Set filter to be used to matching services.
-         */
-        UseServiceBuilder& setFilter(celix::Filter f) { filter = std::move(f); return *this; }
-
-        /**
-         * Sets a optional timeout.
-         * If the timeout is > 0 and there is no matching service, the "build" will block
-         * until a matching service is found or the timeout is expired.
-         *
-         * Note: timeout is only valid if the a single service is used.
-         *
-         */
-        template <typename Rep, typename Period>
-        UseServiceBuilder& setTimeout(std::chrono::duration<Rep, Period> duration) {
-            auto micro =  std::chrono::duration_cast<std::chrono::microseconds>(duration).count();
-            timeoutInSeconds = micro / 1000000;
-            return *this;
-        }
-
-        /**
-         * Adds a use callback function which will be called when the UseServiceBuilder is
-         * "build".
-         *
-         * The use callback function has 1 argument: a reference to the matching service.
-         */
-        UseServiceBuilder& addUseCallback(std::function<void(I&)> cb) {
-            callbacks.emplace_back(std::move(cb));
-            return *this;
-        }
-
-        /**
-         * Adds a use callback function which will be called when the UseServiceBuilder is
-         * "build".
-         *
-         * The use callback function has 2 arguments:
-         *  - A reference to the matching service.
-         *  - A const reference to the matching service properties.
-         */
-        UseServiceBuilder& addUseCallback(std::function<void(I&, const celix::Properties&)> cb) {
-            callbacksWithProperties.emplace_back(std::move(cb));
-            return *this;
-        }
-
-        /**
-         * Adds a use callback function which will be called when the UseServiceBuilder is
-         * "build".
-         *
-         * The use callback function has 3 arguments:
-         *  - A reference to the matching service.
-         *  - A const reference to the matching service properties.
-         *  - A const reference to the bundle owning the matching service.
-         */
-        UseServiceBuilder& addUseCallback(std::function<void(I&, const celix::Properties&, const celix::Bundle&)> cb) {
-            callbacksWithOwner.emplace_back(std::move(cb));
-            return *this;
-        }
-
-        /**
-         * "Builds" the UseServiceBuild and returns directly if no matching service is found
-         * or blocks until:
-         *  - All the use callback functions are called with the highest ranking matching service
-         *  - The timeout has expired
-         *
-         * \returns the number of service called (1 or 0).
-         */
-        std::size_t build() {
-            celix_service_use_options_t opts{};
-            opts.filter.serviceName = name.empty() ? nullptr : name.c_str();
-            opts.filter.ignoreServiceLanguage = true;
-            opts.filter.filter = filter.empty() ? nullptr : filter.getFilterCString();
-            opts.filter.versionRange = versionRange.empty() ? nullptr : versionRange.c_str();
-            opts.waitTimeoutInSeconds = timeoutInSeconds;
-            opts.callbackHandle = this;
-            opts.useWithOwner = [](void* data, void *voidSvc, const celix_properties_t* cProps, const celix_bundle_t* cBnd) {
-                auto* builder = static_cast<UseServiceBuilder<I>*>(data);
-                auto* svc = static_cast<I*>(voidSvc);
-                const Bundle bnd = Bundle{const_cast<celix_bundle_t*>(cBnd)};
-                auto props = celix::Properties::wrap(cProps);
-                for (const auto& func : builder->callbacks) {
-                    func(*svc);
-                }
-                for (const auto& func : builder->callbacksWithProperties) {
-                    func(*svc, *props);
-                }
-                for (const auto& func : builder->callbacksWithOwner) {
-                    func(*svc, *props, bnd);
-                }
-            };
-
-            if (useSingleService) {
-                bool called = celix_bundleContext_useServiceWithOptions(cCtx.get(), &opts);
-                return called ? 1 : 0;
-            } else {
-                return celix_bundleContext_useServicesWithOptions(cCtx.get(), &opts);
-            }
-        }
-    private:
-        const std::shared_ptr<celix_bundle_context_t> cCtx;
-        const std::string name;
-        const bool useSingleService;
-        double timeoutInSeconds{0};
-        celix::Filter filter{};
-        std::string versionRange{};
-        std::vector<std::function<void(I&)>> callbacks{};
-        std::vector<std::function<void(I&, const celix::Properties&)>> callbacksWithProperties{};
-        std::vector<std::function<void(I&, const celix::Properties&, const celix::Bundle&)>> callbacksWithOwner{};
-    };
 
     /**
      * Fluent builder API to track services of type I.
@@ -607,7 +309,7 @@ namespace celix {
         explicit MetaTrackerBuilder(std::shared_ptr<celix_bundle_context_t> _cCtx, std::string _serviceName) :
                 cCtx{std::move(_cCtx)},
                 serviceName{std::move(_serviceName)}
-                {}
+        {}
 
         MetaTrackerBuilder &operator=(MetaTrackerBuilder &&) = delete;
         MetaTrackerBuilder(const MetaTrackerBuilder &) = delete;
